Rename *Instrumenters to *Singletons
