+++ instrumentation/apache-httpclient/apache-httpclient-5.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachehttpclient/v5_0/ApacheHttpClientSingletons.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.apachehttpclient.v4_0;
+package io.opentelemetry.javaagent.instrumentation.apachehttpclient.v5_0;
 
 import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
@@ -13,27 +13,27 @@ import io.opentelemetry.instrumentation.api.instrumenter.http.HttpAttributesExtr
 import io.opentelemetry.instrumentation.api.instrumenter.http.HttpSpanNameExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.http.HttpSpanStatusExtractor;
 import io.opentelemetry.javaagent.instrumentation.api.instrumenter.PeerServiceAttributesExtractor;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.methods.HttpUriRequest;
+import org.apache.hc.core5.http.ClassicHttpRequest;
+import org.apache.hc.core5.http.HttpResponse;
 
-public final class ApacheHttpClientInstrumenters {
+public final class ApacheHttpClientSingletons {
   private static final String INSTRUMENTATION_NAME =
-      "io.opentelemetry.javaagent.apache-httpclient-4.0";
+      "io.opentelemetry.javaagent.apache-httpclient-5.0";
 
-  private static final Instrumenter<HttpUriRequest, HttpResponse> INSTRUMENTER;
+  private static final Instrumenter<ClassicHttpRequest, HttpResponse> INSTRUMENTER;
 
   static {
-    HttpAttributesExtractor<HttpUriRequest, HttpResponse> httpAttributesExtractor =
+    HttpAttributesExtractor<ClassicHttpRequest, HttpResponse> httpAttributesExtractor =
         new ApacheHttpClientHttpAttributesExtractor();
-    SpanNameExtractor<? super HttpUriRequest> spanNameExtractor =
+    SpanNameExtractor<? super ClassicHttpRequest> spanNameExtractor =
         HttpSpanNameExtractor.create(httpAttributesExtractor);
-    SpanStatusExtractor<? super HttpUriRequest, ? super HttpResponse> spanStatusExtractor =
+    SpanStatusExtractor<? super ClassicHttpRequest, ? super HttpResponse> spanStatusExtractor =
         HttpSpanStatusExtractor.create(httpAttributesExtractor);
     ApacheHttpClientNetAttributesExtractor netAttributesExtractor =
         new ApacheHttpClientNetAttributesExtractor();
 
     INSTRUMENTER =
-        Instrumenter.<HttpUriRequest, HttpResponse>newBuilder(
+        Instrumenter.<ClassicHttpRequest, HttpResponse>newBuilder(
                 GlobalOpenTelemetry.get(), INSTRUMENTATION_NAME, spanNameExtractor)
             .setSpanStatusExtractor(spanStatusExtractor)
             .addAttributesExtractor(httpAttributesExtractor)
@@ -42,9 +42,9 @@ public final class ApacheHttpClientInstrumenters {
             .newClientInstrumenter(new HttpHeaderSetter());
   }
 
-  public static Instrumenter<HttpUriRequest, HttpResponse> instrumenter() {
+  public static Instrumenter<ClassicHttpRequest, HttpResponse> instrumenter() {
     return INSTRUMENTER;
   }
 
-  private ApacheHttpClientInstrumenters() {}
+  private ApacheHttpClientSingletons() {}
 }
