+++ instrumentation/armeria-1.3/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/armeria/v1_3/ArmeriaServerBuilderInstrumentation.java
@@ -35,7 +35,7 @@ public class ArmeriaServerBuilderInstrumentation implements TypeInstrumentation 
 
     @Advice.OnMethodEnter
     public static void onEnter(@Advice.This ServerBuilder builder) {
-      builder.decorator(ArmeriaDecorators.SERVER_DECORATOR);
+      builder.decorator(ArmeriaSingletons.SERVER_DECORATOR);
     }
   }
 }
