+++ instrumentation/jedis/jedis-1.4/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v1_4/JedisSingletons.java
@@ -13,7 +13,7 @@ import io.opentelemetry.instrumentation.api.instrumenter.db.DbAttributesExtracto
 import io.opentelemetry.instrumentation.api.instrumenter.db.DbSpanNameExtractor;
 import io.opentelemetry.javaagent.instrumentation.api.instrumenter.PeerServiceAttributesExtractor;
 
-public final class JedisInstrumenters {
+public final class JedisSingletons {
   private static final String INSTRUMENTATION_NAME = "io.opentelemetry.javaagent.jedis-1.4";
 
   private static final Instrumenter<JedisRequest, Void> INSTRUMENTER;
@@ -37,5 +37,5 @@ public final class JedisInstrumenters {
     return INSTRUMENTER;
   }
 
-  private JedisInstrumenters() {}
+  private JedisSingletons() {}
 }
